requirements.txt
aiogram==3.10.0
openai==1.35.13
python-dotenv==1.0.1